requirements.txt
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coverage==5.4
coveralls==3.0.0
cryptography==3.4.6
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
docopt==0.6.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
pep8==1.7.1
phonenumbers==8.12.18
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
PyJWT==2.0.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
